requirements.txt
tensorflow==2.19.0
keras==3.4.1
numpy==1.26.4
ml-dtypes==0.2.0
streamlit==1.29.0
pillow==9.5.0
pandas==1.5.3
protobuf==3.20.3